+++ .test/.archive/.scripts/headless.js
@@ -1,0 +1,3 @@
+// Script - How to Change Your Neocities Thumbnail from ARandomSite: https://arandomsite.neocities.org/writing/#custom-neocities-thumbnail //
+
+const urlParams = new URLSearchParams(window.location.search); if (navigator.userAgent.includes('Screenjesus')) { window.location.replace("/thumbnail/"); }