+++ changelog.md
@@ -1,3 +1,24 @@
 # Changelog
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+
+- Update sitemap
+29 mar 2026, 14:11
+- Add script to check dead links
+29 mar 2026, 14:08
+- Update RSS feeds
+29 mar 2026, 13:20
+- Change blog page url
+29 mar 2026, 13:20
+- Add new sitemap
+29 mar 2026, 13:01
+- Add script to generate sitemaps
+29 mar 2026, 13:01
+- Renamed file to sitemap.old
+29 mar 2026, 13:00
+- Removed test CSS
+29 mar 2026, 13:00
+- Moved unused scripts to a new folder
+29 mar 2026, 13:00
+- Update RSS feed
+29 mar 2026, 11:49